WIP: interpolate_bilinear_lonlat raises IndexError for nans
The `interpolate_bilinear_lonlat` function raises an IndexError
if the longitude or latitude that are passed to it are nans. This
is because some or all of the indices returned by
`bilinear_interpolation_weights` are negative in order to indicate
failure, which may not necessarily be valid indices.

This breaks the HEALPix functions in the `reproject` package.

**Important note**: do not merge this pull request yet. This is
a work in progress. It illustrates the bug by adding a unit test,
but does not fix the bug.