Fix docstrings for 32 bit architectures
We return HEALPix indices as `np.int64`, which is the Numpy index
type on 64-bit but not 32-bit platforms. On 32-bit platforms,
Numpy will add `dtype=int64` to string representations of these
arrays, which causes some doctests to fail.

This workaround calls `print()`, which always prints just the
members of the array.

Fixes #81.