Unpin Numpy version for building wheels
Allow using any version of Numpy between 1.25 and 2. This will
allow us to use the latest release of Numpy on PyPI and take
advantage of prebuilt wheels for more platforms and architectures.

Use the new `NPY_TARGET_VERSION` macro to peform a
backward-compatible supporting versions of Numpy as old as 1.19.