Check interpolate_bilinear_lonlat inputs for invalid values
This is found to fix abort errors that occurred on macOS x86_64
when running the unit tests for the Conda packages. See
https://github.com/conda-forge/astropy-healpix-feedstock/pull/41.

This might indicate that some of the Conda builds have assertions
disabled via the `NDEBUG` preprocessor macro, and some have them
enabled, inconsistently.

Without this patch, any number of assertions should have been
tripped. For example, this one:
https://github.com/astropy/astropy-healpix/blob/v1.1.1/cextern/astrometry.net/healpix.c#L742